relationshionship class

+++ src/Controller/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\User;
+use App\Form\UserType;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -30,16 +31,10 @@ final class UserController extends AbstractController
     #[Route('/user/new', name: 'app_user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+
         $user = new User();
 
-        $form = $this->createFormBuilder($user)
-            ->add('username', TextType::class)
-            ->add('email', EmailType::class)
-            ->add('password', PasswordType::class)
-            ->add('submit', SubmitType::class, [
-                'label' => 'Enregistrer'
-            ])
-            ->getForm();
+        $form = $this->createForm(UserType::class, $user);
 
         $form->handleRequest($request);
 
@@ -51,7 +46,7 @@ final class UserController extends AbstractController
         }
 
         return $this->render('user/new.html.twig', [
-            'form' => $form->createView()
+            'form' => $form
         ]);
     }
 
@@ -66,14 +61,8 @@ final class UserController extends AbstractController
     #[Route('/user/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(User $user, Request $request, EntityManagerInterface $entityManager): Response
     {
-        $form = $this->createFormBuilder($user)
-            ->add('username', TextType::class)
-            ->add('email', EmailType::class)
-            ->add('password', PasswordType::class)
-            ->add('submit', SubmitType::class, [
-                'label' => 'Modifier'
-            ])
-            ->getForm();
+        $form = $this->createForm(UserType::class, $user);
+        $form->remove('password');
 
         $form->handleRequest($request);
 
@@ -84,7 +73,7 @@ final class UserController extends AbstractController
         }
 
         return $this->render('user/edit.html.twig', [
-            'form' => $form->createView(),
+            'form' => $form,
             'user' => $user
         ]);
     }
@@ -92,8 +81,10 @@ final class UserController extends AbstractController
     #[Route('/user/{id}/delete', name: 'app_user_delete', methods: ['POST'])]
     public function delete(User $user, Request $request, EntityManagerInterface $entityManager): Response
     {
-        $entityManager->remove($user);
-        $entityManager->flush();
+        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+            $entityManager->remove($user);
+            $entityManager->flush();
+        }
 
         return $this->redirectToRoute('app_user_index');
     }
